cleaned up current
